corrigindo erro de base duplicada
