+++ database.sql
@@ -1,0 +1,19 @@
+  CREATE TABLE EXAMS (
+    id SERIAL PRIMARY KEY,
+    cpf VARCHAR(14) NOT NULL,
+    nome_paciente VARCHAR(150) NOT NULL,
+    email_paciente VARCHAR(150) NOT NULL,
+    data_nascimento_paciente DATE NOT NULL,
+    endereço_rua_paciente VARCHAR(150) NOT NULL,
+    cidade_paciente VARCHAR(55) NOT NULL,
+    estado_patiente VARCHAR(35) NOT NULL,
+    crm_médico VARCHAR(20) NOT NULL,
+    crm_médico_estado VARCHAR(2) NOT NULL,
+    nome_médico VARCHAR(150) NOT NULL,
+    email_médico VARCHAR(150) NOT NULL,
+    token_resultado_exame VARCHAR(10) NOT NULL,
+    data_exame DATE NOT NULL,
+    tipo_exame VARCHAR(50) NOT NULL,
+    limites_tipo_exame VARCHAR(10) NOT NULL,
+    resultado_tipo_exame INTEGER NOT NULL
+  );
