+++ docker-compose.yml
@@ -12,19 +12,15 @@ services:
     environment:
       REDIS_URL: redis://redis:6379/0
     ports:
-      - "3000:3000"
-
+      - 3000:3000
   postgres:
     image: postgres
-    environment:
-      POSTGRES_HOST_AUTH_METHOD: trust
     volumes: 
       - pgdata_rebase_labs:/var/lib/postgresql/data
-    
+    environment:
+      - POSTGRES_HOST_AUTH_METHOD=trust
   redis:
     image: redis
-
-
 volumes:
   rubygems:
   gemconfig:
